requirements.txt
asgiref==3.4.1
boto3==1.24.84
botocore==1.27.84
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
coverage==6.3.1
DateTime==4.3
dj-database-url==0.5.0
Django==3.2.11
django-cleanup==5.2.0
django-cors-headers==3.11.0
django-dbbackup==3.3.0
django-extensions==3.2.0
django-filter==21.1
django-health-check==3.16.5
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
google-api-core==2.4.0
google-api-python-client==2.51.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.32.0
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
gspread==5.1.1
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
Janome==0.4.1
jmespath==1.0.1
oauth2client==4.1.3
oauthlib==3.1.1
packaging==21.3
Pillow==9.0.1
proto-plus==1.19.8
protobuf==3.19.3
psutil==5.9.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
s3transfer==0.6.0
schedule==1.1.0
six==1.16.0
slackweb==1.0.5
sqlparse==0.4.2
stripe==2.65.0
tenacity==8.1.0
uritemplate==4.1.1
urllib3==1.26.8
websocket-client==1.1.0
whitenoise==5.3.0
zope.interface==5.4.0
